Test visibility for food

+++ Assets/Scripts/Datatypes/Coord.cs
@@ -12,7 +12,7 @@ public struct Coord {
         this.y = y;
     }
 
-    public static float SqrDistance (Coord a, Coord b) {
+    public static int SqrDistance (Coord a, Coord b) {
         return (a.x - b.x) * (a.x - b.x) + (a.y - b.y) * (a.y - b.y);
     }
